Refactor imports to consolidate CommandType into libs.types for improved organization and clarity
