+++ integrations/discord/parser.py
@@ -10,7 +10,7 @@ from discord.channel import TextChannel
 import libs.global_value as g
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
 from integrations.protocols import MsgData, PostData, StatusData
-from libs.domain.datamodels import CommandType
+from libs.types import CommandType
 
 if TYPE_CHECKING:
     from integrations.discord.adapter import ServiceAdapter
