+++ integrations/protocols.py
@@ -5,8 +5,7 @@ integrations/protocols.py
 from dataclasses import dataclass, field, fields, is_dataclass
 from typing import TYPE_CHECKING, Any, Optional, Protocol
 
-from libs.domain.datamodels import CommandType
-from libs.types import ActionStatus, ChannelType, MessageStatus
+from libs.types import ActionStatus, ChannelType, CommandType, MessageStatus
 
 if TYPE_CHECKING:
     from pathlib import Path  # noqa: F401
