+++ integrations/slack/events/home_tab/versus.py
@@ -10,7 +10,7 @@ from integrations.slack.adapter import ServiceAdapter
 from integrations.slack.events.handler_registry import register
 from integrations.slack.events.home_tab import ui_parts
 from libs.commands.results import versus
-from libs.domain.datamodels import CommandType
+from libs.types import CommandType
 from libs.utils import dictutil
 from libs.utils.timekit import Delimiter, Format
 from libs.utils.timekit import ExtendedDatetime as ExtDt
