+++ integrations/slack/parser.py
@@ -8,8 +8,7 @@ from typing import TYPE_CHECKING, Any, cast
 import libs.global_value as g
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
 from integrations.protocols import MsgData, PostData, StatusData
-from libs.domain.datamodels import CommandType
-from libs.types import ChannelType, MessageStatus
+from libs.types import ChannelType, CommandType, MessageStatus
 
 if TYPE_CHECKING:
     from integrations.slack.adapter import ServiceAdapter
