+++ libs/commands/graph/entry.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 import libs.global_value as g
 from libs.bootstrap.section import SubCommands
 from libs.commands.graph import personal, rating, summary
-from libs.domain.datamodels import CommandType
+from libs.types import CommandType
 from libs.utils import dictutil
 
 if TYPE_CHECKING:
