+++ libs/commands/help/entry.py
@@ -8,8 +8,7 @@ from typing import TYPE_CHECKING
 import libs.global_value as g
 from libs.bootstrap.section import SubCommands
 from libs.data import lookup
-from libs.domain.datamodels import CommandType
-from libs.types import StyleOptions
+from libs.types import CommandType, StyleOptions
 from libs.utils import dictutil
 from libs.utils.timekit import ExtendedDatetime as ExtDt
