+++ libs/commands/registry/member.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, TypedDict, cast
 import libs.global_value as g
 from libs.bootstrap.section import BaseSection
 from libs.data import loader, modify
-from libs.domain.datamodels import CommandType
+from libs.types import CommandType
 from libs.utils import dbutil, textutil, validator
 
 if TYPE_CHECKING:
