+++ libs/functions/message.py
@@ -10,8 +10,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from libs.domain.datamodels import CommandType
 from libs.functions.compose import text_item
+from libs.types import CommandType
 from libs.utils.timekit import ExtendedDatetime as ExtDt
 from libs.utils.timekit import Format
