+++ libs/types.py
@@ -64,6 +64,31 @@ class ChannelType(StrEnum):
     """未定義状態"""
 
 
+class CommandType(StrEnum):
+    """実行(する/した)サブコマンド県設定ファイルセクション名"""
+
+    RESULTS = "results"
+    """成績サマリ"""
+    GRAPH = "graph"
+    """グラフ生成"""
+    RANKING = "ranking"
+    """ランキング"""
+    RATING = "rating"
+    """レーティング"""
+    REPORT = "report"
+    """レポート"""
+    MEMBER_LIST = "member"
+    """メンバー一覧"""
+    TEAM_LIST = "team"
+    """チーム一覧"""
+    HELP = "help"
+    """ヘルプ"""
+    COMPARISON = "comparison"
+    """突合処理"""
+    UNKNOWN = "unknown"
+    """未定義"""
+
+
 class ServiceType(StrEnum):
     """連携先サービス"""
